+++ useScroll/useScroll.js
@@ -172,14 +172,18 @@ const useThemeScroll = (props, instances) => {
 	}
 
 	function focusOnItem () {
+		let isItemFocused = false;
+
 		if (mutableRef.current.indexToFocus !== null && typeof themeScrollContentHandle.current.focusByIndex === 'function') {
 			themeScrollContentHandle.current.focusByIndex(mutableRef.current.indexToFocus);
 			mutableRef.current.indexToFocus = null;
+			isItemFocused = true;
 		}
 
 		if (mutableRef.current.nodeToFocus !== null && typeof themeScrollContentHandle.current.focusOnNode === 'function') {
 			themeScrollContentHandle.current.focusOnNode(mutableRef.current.nodeToFocus);
 			mutableRef.current.nodeToFocus = null;
+			isItemFocused = true;
 		}
 
 		if (mutableRef.current.pointToFocus !== null) {
@@ -196,10 +200,15 @@ const useThemeScroll = (props, instances) => {
 
 				if (target) {
 					Spotlight.focus(target);
+					isItemFocused = true;
 				}
 			}
 
 			mutableRef.current.pointToFocus = null;
+		}
+
+		if (Spotlight.getPointerMode() && !isItemFocused) {
+			Spotlight.focus(scrollContainerRef.current, {enterTo: 'topmost'});
 		}
 	}